requirements.txt
anthropic>=0.35.0
google-generativeai>=0.8.0
boto3>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
fastapi
uvicorn
streamlit
